resume-tailor-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pypdf==4.3.1
python-docx==1.1.2
google-generativeai==0.8.3
groq==0.11.0
openai==1.51.0
mistralai==1.2.5
sqlalchemy==2.0.35
python-multipart==0.0.12
aiofiles==24.1.0
slowapi==0.1.9
alembic==1.13.3
google-api-python-client==2.153.0
google-auth==2.36.0
